feat(backend): Add global CORS configuration middleware
